scripts/requirements_index.txt
azure-search-documents==11.4.0
azure-identity==1.15.0
# Add other dependencies if setup_search_index.py requires them (e.g., for reading JSON)

